# specs/001-todo-app/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Todo API
  description: API for managing todo tasks
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.example.com
    description: Production server
paths:
  /tasks:
    get:
      summary: Get all tasks
      description: Retrieve a list of all todo tasks
      responses:
        '200':
          description: A list of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
    post:
      summary: Create a new task
      description: Create a new todo task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreateRequest'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /tasks/{id}:
    get:
      summary: Get a task by ID
      description: Retrieve a specific task by its ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Task found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Update a task
      description: Update an existing task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdateRequest'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete a task
      description: Delete a specific task by its ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '204':
          description: Task deleted successfully
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    Task:
      type: object
      required:
        - id
        - title
        - status
        - created_at
      properties:
        id:
          type: integer
          example: 1
          description: Unique identifier for the task
        title:
          type: string
          example: "Buy groceries"
          description: The task title
          minLength: 1
          maxLength: 255
        description:
          type: string
          example: "Milk, bread, eggs"
          description: Extended task details
          maxLength: 1000
        status:
          type: string
          enum: [PENDING, COMPLETED]
          example: "PENDING"
          description: Current status of the task
        created_at:
          type: string
          format: date-time
          example: "2023-10-05T14:48:00Z"
          description: Timestamp when the task was created
    TaskCreateRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          example: "Buy groceries"
          description: The task title
          minLength: 1
          maxLength: 255
        description:
          type: string
          example: "Milk, bread, eggs"
          description: Extended task details
          maxLength: 1000
        status:
          type: string
          enum: [PENDING, COMPLETED]
          example: "PENDING"
          description: Initial status of the task, defaults to PENDING
    TaskUpdateRequest:
      type: object
      properties:
        title:
          type: string
          example: "Buy groceries"
          description: The task title
          minLength: 1
          maxLength: 255
        description:
          type: string
          example: "Milk, bread, eggs"
          description: Extended task details
          maxLength: 1000
        status:
          type: string
          enum: [PENDING, COMPLETED]
          example: "COMPLETED"
          description: Updated status of the task
    Error:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          example: "Task not found"
          description: Error message